+++ src/main/java/ru/yandex/practicum/filmorate/model/Counter.java
@@ -1,9 +1,9 @@
 package ru.yandex.practicum.filmorate.model;
 
 public class Counter {
-    public int counter = 1;
+    public long counter = 1;
 
-    public int count() {
+    public long count() {
         return counter++;
     }
 }
